+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 # 0.3.0-alpha.3 [unreleased]
-- refactor/opt: Expanded UninitializedIpfs functionality [PR 21]
 - feat/upnp: Use libp2p-nat for handling port forwarding [PR 23]
+
 [PR 23]: https://github.com/dariusc93/rust-ipfs/pull/23
-[PR 21]: https://github.com/dariusc93/rust-ipfs/pull/21
 
 # 0.3.0-alpha.2
 - fix: Poll receiving oneshot channel directly [PR 20]